feat: Ceremony 도메인 파일 업로드 분리
