+++ app-main/src/main/java/net/causw/app/main/domain/community/ceremony/api/v2/controller/CeremonyController.java
@@ -1,20 +1,16 @@
 package net.causw.app.main.domain.community.ceremony.api.v2.controller;
 
-import java.util.List;
-
 import org.springframework.data.domain.Page;
 import org.springframework.http.HttpStatus;
-import org.springframework.http.MediaType;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.bind.annotation.RequestPart;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
-import org.springframework.web.multipart.MultipartFile;
 
 import net.causw.app.main.domain.community.ceremony.api.v2.dto.request.CeremonyCreateRequest;
 import net.causw.app.main.domain.community.ceremony.api.v2.dto.response.CeremonyDetailResponse;
@@ -43,15 +39,14 @@ public class CeremonyController {
 	private final CeremonyDtoMapper ceremonyDtoMapper;
 	private final CeremonyService ceremonyService;
 
-	@PostMapping(consumes = MediaType.MULTIPART_FORM_DATA_VALUE, produces = MediaType.APPLICATION_JSON_VALUE)
+	@PostMapping
 	@ResponseStatus(value = HttpStatus.CREATED)
 	@Operation(summary = "사용자 본인의 경조사 생성", description = "사용자 본인의 경조사 생성합니다.")
 	public ApiResponse<CeremonyDetailResponse> createCeremony(
 		@AuthenticationPrincipal CustomUserDetails userDetails,
-		@RequestPart(value = "ceremonyCreateRequest") @Valid CeremonyCreateRequest request,
-		@RequestPart(value = "imageFileList", required = false) List<MultipartFile> imageFileList) {
+		@RequestBody @Valid CeremonyCreateRequest request) {
 		CeremonyCreateCommand command = ceremonyDtoMapper.toCreateCommand(request);
-		CeremonyDetailResult result = ceremonyService.createCeremony(userDetails.getUser(), command, imageFileList);
+		CeremonyDetailResult result = ceremonyService.createCeremony(userDetails.getUser(), command);
 		return ApiResponse.success(ceremonyDtoMapper.toDetailResponse(result));
 	}
