+++ app-main/src/main/java/net/causw/app/main/domain/community/ceremony/service/dto/request/CeremonyCreateCommand.java
@@ -28,5 +28,6 @@ public record CeremonyCreateCommand(
 	String contact,
 	String link,
 	Boolean isSetAll,
-	List<Integer> targetAdmissionYears) {
+	List<Integer> targetAdmissionYears,
+	List<String> imageUuids) {
 }
